+++ r-naming.tf
@@ -1,4 +1,4 @@
-resource "azurecaf_name" "appgw" {
+data "azurecaf_name" "appgw" {
   name          = var.stack
   resource_type = "azurerm_application_gateway"
   prefixes      = var.name_prefix == "" ? null : [local.name_prefix]
@@ -8,7 +8,7 @@ resource "azurecaf_name" "appgw" {
   separator     = "-"
 }
 
-resource "azurecaf_name" "subnet_appgw" {
+data "azurecaf_name" "subnet_appgw" {
   name          = var.stack
   resource_type = "azurerm_subnet"
   prefixes      = var.name_prefix == "" ? null : [local.name_prefix]
@@ -18,7 +18,7 @@ resource "azurecaf_name" "subnet_appgw" {
   separator     = "-"
 }
 
-resource "azurecaf_name" "nsg_appgw" {
+data "azurecaf_name" "nsg_appgw" {
   name          = var.stack
   resource_type = "azurerm_network_security_group"
   prefixes      = var.name_prefix == "" ? null : [local.name_prefix]
@@ -28,7 +28,7 @@ resource "azurecaf_name" "nsg_appgw" {
   separator     = "-"
 }
 
-resource "azurecaf_name" "pip_appgw" {
+data "azurecaf_name" "pip_appgw" {
   name          = var.stack
   resource_type = "azurerm_public_ip"
   prefixes      = var.name_prefix == "" ? null : [local.name_prefix]
@@ -38,7 +38,7 @@ resource "azurecaf_name" "pip_appgw" {
   separator     = "-"
 }
 
-resource "azurecaf_name" "nsr_https" {
+data "azurecaf_name" "nsr_https" {
   name          = var.stack
   resource_type = "azurerm_network_security_group_rule"
   prefixes      = var.name_prefix == "" ? null : [local.name_prefix]
@@ -48,7 +48,7 @@ resource "azurecaf_name" "nsr_https" {
   separator     = "-"
 }
 
-resource "azurecaf_name" "nsr_healthcheck" {
+data "azurecaf_name" "nsr_healthcheck" {
   name          = var.stack
   resource_type = "azurerm_network_security_group_rule"
   prefixes      = var.name_prefix == "" ? null : [local.name_prefix]
@@ -58,7 +58,7 @@ resource "azurecaf_name" "nsr_healthcheck" {
   separator     = "-"
 }
 
-resource "azurecaf_name" "frontipconfig" {
+data "azurecaf_name" "frontipconfig" {
   name          = var.stack
   resource_type = "azurerm_public_ip"
   prefixes      = compact([var.use_caf_naming ? "frontipconfig" : "", local.name_prefix])
@@ -68,7 +68,7 @@ resource "azurecaf_name" "frontipconfig" {
   separator     = "-"
 }
 
-resource "azurecaf_name" "frontipconfig_priv" {
+data "azurecaf_name" "frontipconfig_priv" {
   name          = var.stack
   resource_type = "azurerm_public_ip"
   prefixes      = compact([var.use_caf_naming ? "frontipconfig-priv" : "", local.name_prefix])
@@ -78,7 +78,7 @@ resource "azurecaf_name" "frontipconfig_priv" {
   separator     = "-"
 }
 
-resource "azurecaf_name" "gwipconfig" {
+data "azurecaf_name" "gwipconfig" {
   name          = var.stack
   resource_type = "azurerm_public_ip"
   prefixes      = compact([var.use_caf_naming ? "gwipconfig" : "", local.name_prefix])